+++ src/Extensions/Syndication.php
@@ -31,7 +31,7 @@ final class Syndication extends BaseExtension
     {
         return [
             self::ELEMENT_UPDATE_PERIOD => UpdatePeriod::class,
-            self::ELEMENT_UPDATE_FREQUENCY => "int",
+            self::ELEMENT_UPDATE_FREQUENCY => "positive-int",
             self::ELEMENT_UPDATE_BASE => "string",
         ];
     }
@@ -39,9 +39,5 @@ final class Syndication extends BaseExtension
     public function configureChannelOptions(OptionsResolver $resolver, Generator $generator): void
     {
         $this->registerElements($resolver);
-        $resolver->setAllowedValues(
-            $this->getElementName(self::ELEMENT_UPDATE_FREQUENCY),
-            static fn(int $value): bool => ($value >= 1)
-        );
     }
 }